Minor refactor to support booking fixes.
