+++ src/main/java/org/dcsa/core/events/service/impl/LocationServiceImpl.java
@@ -5,7 +5,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.dcsa.core.events.model.Address;
 import org.dcsa.core.events.model.Facility;
 import org.dcsa.core.events.model.Location;
-import org.dcsa.core.events.model.mappers.LocationMapper;
+import org.dcsa.core.events.model.mapper.LocationMapper;
 import org.dcsa.core.events.model.transferobjects.LocationTO;
 import org.dcsa.core.events.repository.LocationRepository;
 import org.dcsa.core.events.repository.UnLocationRepository;
